// src/app/models/Usuarios.ts
import { Roles } from "./Role";

export class Usuario {
    idUsuario: number=0;
    usNombre: string="";
    usApellido: string="";
    usFecNacimiento:Date=new Date();
    usCorreo: string="";
    username: string="";
    password: string="";
    usEnable: boolean=false;
    role:Roles = new Roles();
}